frontend
Geradores antigos colocados na pasta old

Novos geradores criados (vazios) dentro de suas respectivas pastas
Seguindo a estrutura colocada como comentário no src/cli/frontend/vue-vite/src/generate.ts
