+++ src/cli/frontend/vue-vite/src/api/generate.ts
@@ -1,5 +1,6 @@
 import { Model } from "../../../../../language/generated/ast.js";
 
+
 export function generate(model: Model, target_folder: string) : void {
 
 }
