+++ src/cli/frontend/vue-vite/src/assets/generate.ts
