+++ src/cli/frontend/vue-vite/src/modules/generate.ts
