+++ src/cli/frontend/vue-vite/src/modules/routes/generate.ts
@@ -1,4 +1,4 @@
-import { Model } from "../../../../../language/generated/ast.js";
+import { Model } from "../../../../../../language/generated/ast.js";
 
 export function generate(model: Model, target_folder: string) : void {
