+++ src/cli/frontend/vue-vite/src/old/composition_old/generate.ts
