+++ src/cli/frontend/vue-vite/src/old/layouts_old/full/horizontal-header/generate.ts
