+++ src/cli/frontend/vue-vite/src/old/layouts_old/full/horizontal-sidebar/generate.ts
