+++ src/cli/frontend/vue-vite/src/old/layouts_old/full/logo/generate.ts
