+++ src/cli/frontend/vue-vite/src/old/layouts_old/full/vertical-header/generate.ts
