+++ src/cli/frontend/vue-vite/src/old/layouts_old/full/vertical-sidebar/generate.ts
