+++ src/cli/frontend/vue-vite/src/old/router_old/generate.ts
