+++ src/cli/frontend/vue-vite/src/old/scss_old/components/generate.ts
