+++ src/cli/frontend/vue-vite/src/old/scss_old/pages/generate.ts
