+++ src/cli/frontend/vue-vite/src/old/services_old/generate.ts
