+++ src/cli/frontend/vue-vite/src/old/services_old/requires/generate.ts
