+++ src/cli/frontend/vue-vite/src/old/stores_old/apps/userprofile/generate.ts
