+++ src/cli/frontend/vue-vite/src/old/theme_old/generate.ts
