+++ src/cli/frontend/vue-vite/src/old/views_old/authentication/generate.ts
