+++ src/cli/frontend/vue-vite/src/old/views_old/model/generate.ts
