+++ src/cli/frontend/vue-vite/src/routes/generate.ts
